Updated the portion of fill_in_indtargs that looks through the data segments of a program to find pointers such that 64-bit code needs to find a 64-bit pointer, otherwise it's likely not a pointer.  Was getting spurious false hits from an array of shorts that contained characters such that strings like 00490078 would be found as a valid pointer, when in fact it isn't.
Former-commit-id: 0a8b679c83422264fcdf91b8155fe4cd6e9ddc81

+++ libIRDB/test/check_thunks.cpp
@@ -124,7 +124,7 @@ void check_func_for_thunk_offsets(Function_t *func, Instruction_t* thunk_insn,
 	string reg, string offset)
 {
 
-	bool possible_target(int p, uintptr_t at=0);
+	bool possible_target(uintptr_t p, uintptr_t at=0);
 
 
 	int thunk_base=thunk_insn->GetFallthrough()->GetAddress()->GetVirtualOffset()+
