+++ libIRDB/test/check_thunks.hpp
@@ -29,7 +29,7 @@ void find_all_module_starts(libIRDB::FileIR_t* firp, std::set<int>& thunk_bases)
 void check_for_thunks(libIRDB::FileIR_t* firp, const std::set<int>& thunk_bases);
 
 
-bool possible_target(int p, uintptr_t at=0);
+bool possible_target(uintptr_t p, uintptr_t at=0);
 
 
 #endif
